Open now calls the plugin

+++ gui/scroom/src/view.cc
@@ -200,7 +200,7 @@ void View::updateZoom()
 
     int zMax = MaxZoom - minZoom;
     zMax = std::max(zMax, 1+MaxZoom-zoom);
-    zMax = std::min((unsigned int)zMax, sizeof(zoomfactor)/sizeof(zoomfactor[0]));
+    zMax = std::min((size_t)zMax, sizeof(zoomfactor)/sizeof(zoomfactor[0]));
     bool zoomFound = false;
   
     gtk_list_store_clear(zoomItems);
